habit-tracker/requirements.txt
pygame==2.5.2
Pillow==10.2.0
pytest==8.0.0
pytest-cov==4.1.0
